feat(edge): show correct scripts
